Added MatchAlgoExample.py & MatchFunction.py
I made a slightly more accessible version of the code I used to make those graphs I showed you all last week... MatchAlgoExample.py is mostly comments and just calls a function or two from MatchFunction.py which implements the basic algorithm of 1. pick random stock charts all of the same length as the pattern you input, 2. fit them all to the 'target' stock you want to forecast for 3. Make a forecast by taking a weighted average of the stock charts which fit the target best.

Later on - Sprint 3 - some of us will want to play with inputs to the function (match()) making the forecast to see if we can improve performance. It takes as input a weighting function and a pick_comparables function. Use help(match) once you have it running for more details. I think we'll need to add an 'industry' column to our data before picking stocks to use in the forecast gets interesting (now its just the x stocks with the lowest squared error w/ respect to target stock).